+++ src/main/kotlin/no/skatteetaten/aurora/gobo/graphql/toxiproxy/ToxiProxyToxicMutation.kt
@@ -135,7 +135,7 @@ data class ToxicInput(
     val name: String,
     val type: String,
     val stream: String,
-    val toxicity: Int,
+    val toxicity: Double,
     val attributes: List<ToxicAttributeInput>
 )
